+++ client/src/shared/components/HeaderElements/TopBar.js
@@ -1,0 +1,34 @@
+import React from "react";
+
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+
+import { makeStyles } from "@material-ui/core/styles";
+
+const drawerWidth = "15rem";
+
+const useStyles = makeStyles(theme => ({
+  appBar: {
+    [theme.breakpoints.up("sm")]: {
+      width: `calc(100% - ${drawerWidth})`,
+      marginLeft: drawerWidth
+    }
+  },
+  toolbar: {
+    display: 'flex',
+    justifyContent: 'flex-end',
+    backgroundColor: '#fafafa',
+    color: 'initial'
+  }
+}));
+
+export function TopBar(props) {
+  const classes = useStyles();
+  return (
+    <AppBar className={classes.appBar}>
+      <Toolbar className={classes.toolbar}>
+        {props.children}
+      </Toolbar>
+    </AppBar>
+  )
+}
